% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_1_utils.R
\name{get_msmt_inds}
\alias{get_msmt_inds}
\title{Return a tibble, decomposing MSMT variable name into a table, row and column indices}
\usage{
get_msmt_inds(
  x,
  temp_expr = "^(r|R)(.{2})([0-9]{2}[A-Za-z]{0,1})([0-9]{1,2}[A-Za-z]{0,1})$"
)
}
\arguments{
\item{x}{a single character or a character vector, containing MSMT variable name(s)}

\item{temp_expr}{a regex expression, containing MSMT variable name format}
}
\description{
Return a tibble, decomposing MSMT variable name into a table, row and column indices
}
