% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_1_utils.R
\name{is_long}
\alias{is_long}
\title{Indicate, whether a tibble containing MSMT is in a long or a wide format}
\usage{
is_long(data, id_vec)
}
\arguments{
\item{data}{A tibble contining MSMT data}

\item{id_vec}{A character vector containing variable names used for matching cases}
}
\description{
This function can be used to determine the optimal mode of joining MSMT data tibbles.
For these purposes, data is considered to be long, whennever it either contains
more than one instance of every observed combination of indicator values per or, or
when at least one of the selected indicators is missing in the data.
}
