% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_1_utils.R
\name{is_msmt_var}
\alias{is_msmt_var}
\title{Does a variable name match MSMT data name format?}
\usage{
is_msmt_var(
  x,
  temp_expr = "^(r|R)(.{2})([0-9]{2}[A-Za-z]{0,1})([0-9]{1,2}[A-Za-z]{0,1})$"
)
}
\arguments{
\item{x}{a single character or a character vector, containing the variable name(s)}

\item{temp_expr}{a regex expression, containing MSMT variable name format}
}
\description{
Does a variable name match MSMT data name format?
}
