% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_2_1_msmt_addresses.R
\name{msmt_addresses}
\alias{msmt_addresses}
\title{Download and prepare data with school addresses
This function downloads data from the address book of all school and educational facilities in Czech Republic (Rejstříku škol a školských zařízení)}
\usage{
msmt_addresses(NUTS3 = "all")
}
\arguments{
\item{NUTS3}{A character vector, containing either selected NUTS3 region identifiers, or a single character 'all' (default) for the entire country.}
}
\value{
A tibble, containing addresses and other information about schools and educational facilities, on the level of p_izo
}
\description{
Download and prepare data with school addresses
This function downloads data from the address book of all school and educational facilities in Czech Republic (Rejstříku škol a školských zařízení)
}
\examples{
data_addresses_PRAGUE <- msmt_addresses(NUTS3 = "CZ010")

}
