% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_2_5_msmt_distances.R
\name{msmt_distances}
\alias{msmt_distances}
\title{Compute distances between locations
This function takes as input coordinates and identificators and returns a distance matrix that can be used in further functions.}
\usage{
msmt_distances(id, X, Y, id_row = NULL, id_col = NULL)
}
\arguments{
\item{id}{A character vector containing ID values of all used locations}

\item{X}{A numeric vector of the same length, indicating X coordinates}

\item{Y}{A numeric vector of the same length, indicating Y coordinates}

\item{id_row}{Optional - a subset of id to constitute the rows of distance matrix}

\item{id_col}{Optional - a subset of id to constitute the columns of distance matrix}
}
\value{
A unit matrix containing pairwise distances between locations in meters, with id as row and column names.
The units can be removed by using units::drop_units()
}
\description{
Compute distances between locations
This function takes as input coordinates and identificators and returns a distance matrix that can be used in further functions.
}
\examples{
data_addresses <- msmt_addresses(NUTS3 = "CZ010")
data_addresses_wcoord <- msmt_coordinates(data_addresses)
data_addresses_wcoord <- na.omit(data_addresses_wcoord[,c("red_izo", "RedRUAINKod_geometry")])
id <- data_addresses_wcoord$red_izo
X <- unlist(data_addresses_wcoord$RedRUAINKod_geometry)[c(TRUE,FALSE)]
Y <- unlist(data_addresses_wcoord$RedRUAINKod_geometry)[c(FALSE,TRUE)]
dmat <- msmt_distances(id = id, X = X, Y = Y)

}
