% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_2_msmt_forms.R
\name{msmt_forms}
\alias{msmt_forms}
\title{A function to browse and select MSMT datasets}
\usage{
msmt_forms(
  what = "forms",
  language = "cs",
  search_string = NULL,
  exact_match = TRUE,
  ...
)
}
\arguments{
\item{what}{What information would you like to receive. Defaults to "forms". See details for more}

\item{language}{Language of output, currently available are Czech original ("cs") and english translation from DeepL ("en")}

\item{search_string}{Optional. A perl regular expression containing a search term to use}

\item{exact_match}{Should an exact match, or an approximate match be used if parameter \code{what = "search"}. Defaults to true.}

\item{...}{Additional arguments that can be passed to \code{agrepl()} if \code{exact_match = FALSE}.}
}
\value{
A tibble with columns form (form codes)
}
\description{
A function to browse and select MSMT datasets
}
\details{
\describe{
\item{what}{takes single character input, indicating, which information to return. Available options are:}
\describe{
\item{forms}{A tibble containing form codes, numbers, names, the year of the first and the currently last available form.}
\item{tables}{A nested named list, containing the reconstruction of form tables. This might be useful for interpreting some variable tables.}
\item{browser}{A nested named list divided into forms, tables in forms and variables in tables. Names in the list represent form names, illustrative first variable label of each table and variable labels inside the table. Elements of the list are variable names. This can be useful for brosing available variables using the '$' operator.}
\item{labels}{A nested named list of variable labels, divided into forms. The names in the list represent form codes and variable names. Can be used for labelling variables.}
\item{availability}{A tibble, showing detailed form availability in years.}
\item{search_string}{Used in combination with \code{search_string} parameter. Returns a tibble with instances of variables and their forms, matching the search term.}
}
}
}
\examples{
\dontrun{form_tibble <- msmt_forms(what = "search", search_string = "znevýhodnění")}

}
