% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_2_3_msmt_gender_names.R
\name{msmt_gender_names}
\alias{msmt_gender_names}
\title{Guess gender from name}
\usage{
msmt_gender_names(
  names_vec,
  match_char = "á$",
  countries = c("CZ", "SK", "PL"),
  timeout = 360
)
}
\arguments{
\item{names_vec}{A character vector containing full names (including surnames)}

\item{match_char}{A character containing a regular expression indicating female. Default is a Czech ending letter for surnames "á".
To omit, use "YYYY", or some other regular expression that won't produce any match.}

\item{countries}{Country codes to be used in matching names from the database. Note that adding too many countries might result in a very long computation}

\item{timeout}{Increases timeout for dowloads with options(timeout = timeout) in order to allow the download of the international names database. Defaults to 200, to skip, set value to NULL.}
}
\value{
A tibble matching original names to guessed genders. Also includes variables indicating,
how many names belonging to a gender in a given languages have been matched to the original name.
Also contains a dummy, indicating a match with the match character 'match_char'.
}
\description{
This function guesses the assigned gender of a person from his/her name. Can be used in conjunction with \code{msmt_addressess()}.
}
\details{
To cite the database used in the function, use:
Martínez, G. L., de Juano-i-Ribes, H. S., Yin, D., Le Feuvre, B., Hamdan-Livramento, I., Saito, K., & Raffo, J. (2021). Expanding the World Gender-Name Dictionary: WGND 2.0.
}
\examples{
data_addresses <- msmt_addresses(NUTS3 = "CZ010")
data_genders <- msmt_gender_names(data_addresses$ReditelJmeno)

}
