% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_1_4_msmt_names.R
\name{msmt_names}
\alias{msmt_names}
\title{Replace variable names with intelligible labels}
\usage{
msmt_names(
  data,
  form,
  language = "cs",
  id_vec = c("red_izo", "izo", "p_izo", "year"),
  var_sel = NULL,
  data_labels = NULL
)
}
\arguments{
\item{data}{A tibble with msmt data.}

\item{form}{A single character, specifying which MSMT form the data belong to (see \code{msmt_forms()}).}

\item{language}{Language of output, currently available are Czech original ("cs") and english translation from DeepL ("en")}

\item{id_vec}{A character vector, containing the names of variables used as identifiers. Defaults to c("red_izo", "izo", "p_izo", "year"). Identifiers are not renamed.}

\item{var_sel}{A character vector, containing the names of variables to be renamed. If NULL (default), all variables will be renamed.}

\item{data_labels}{A named character vector, containing the labels to be assigned to variables, named with the respective variable names. If NULL (default), the default labels of \code{MSMT} package will be used.}
}
\value{
A tibble with renamed data
}
\description{
Replace variable names with intelligible labels
}
\examples{
data_locations <- msmt_download_data(form = "M03", years = 2015:2022)
data_list <- msmt_read_data(data_locations = data_locations)
named_tibble <- msmt_names(data = data_list$wide, form = data_list$form)

}
