% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_1_2_msmt_read_data.R
\name{msmt_read_data}
\alias{msmt_read_data}
\title{Reads and combines local copies of MSMT data}
\usage{
msmt_read_data(data_locations, id_vec = c("red_izo", "izo", "p_izo", "year"))
}
\arguments{
\item{data_locations}{Named character vector with file locations of MSMT data in .xlsx format.}

\item{id_vec}{A character vector, containing the names of variables to be used as identifiers. Defaults to c("red_izo", "izo", "p_izo", "year").}
}
\value{
A named list, containing:
\describe{
\item{form}{A single character, indicating MSMT form.}
\item{wide}{A single tibble, containing merged data originally represented in a wide format.}
\item{wide_id}{A single tibble, id data, such as addresses etc., detached from original wide format data.}
\item{long}{A named list, containing separate tibbles of long data. Each tibble represents a separate excel sheet in the original data. The sheets are merged over years.}
\item{row}{A named list, containing separate tibbles of all data. Each tibble represents a separate excel sheet in the original data.}
}
}
\description{
Reads and combines local copies of MSMT data
}
\details{
\describe{
\item{data_locations}{Can be an output of \code{msmt_download_data()}. The characters contain file locations of .xlsx files with MSMT form data.
Names contain the form (F) and year (Y) of the data in the following format FFF_YYYY.
Only read data belonging to one form at a time (e.g. any time range from form S01).}
}
}
\examples{
data_locations <- msmt_download_data(form = "M03", years = 2015:2022)
data_list <- msmt_read_data(data_locations = data_locations)

}
