% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_1_utils.R
\name{msmt_table_names}
\alias{msmt_table_names}
\title{Construct a MSMT table with variable names}
\usage{
msmt_table_names(
  table_ind,
  row_inds,
  col_inds,
  extra_row = NULL,
  extra_col = NULL,
  omits = NULL,
  prefix = "r",
  column_names = NULL,
  rows_names = NULL
)
}
\arguments{
\item{table_ind}{a single numeric or character, index of table, e.g. 1 or "01" or "01B"}

\item{row_inds}{a numeric or a character vector, containing indices of rows, e.g. c(1,2,3)}

\item{col_inds}{a numeric or a character vector, containing indices of columns, e.g. c(1,2,3)}

\item{extra_row}{a character vector, containing indices of special rows, e.g c("07a", "07b")}

\item{extra_col}{a character vector, containing indices of special columns, e.g c("07a", "07b")}

\item{omits}{full variable names that should be ommited (empty cells in a form table)}

\item{prefix}{a prefix for variable names, defaults to "r"}

\item{column_names}{assign column names to output table?}

\item{rows_names}{assign column names to output table?}
}
\value{
A matrix replicating a form table, containing variable names
}
\description{
This function builds a table of variable names that copies the cells of a variable table in MSMT forms.
}
\examples{
msmt_table_names(table_ind = "03",
                 row_inds = 1:7,
                 col_inds = 1:4,
                 extra_row = "07a",
                 omits = c("r03051",
                           "r03052"))

}
