% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_2_4_msmt_titles.R
\name{msmt_titles}
\alias{msmt_titles}
\title{Guess highest level of education from titles included in name
This function guesses the highest level of education from titles recorded in full name (e.g. "Mgr."). Can be used in conjunction with \code{msmt_addressess()}.}
\usage{
msmt_titles(names_vec)
}
\arguments{
\item{names_vec}{A character vector containing full names (inclusing surnames and titles)}
}
\value{
A tibble containing the original name and an ordered variable indication the guessed education, as well as a general idea of the observed title.
}
\description{
Guess highest level of education from titles included in name
This function guesses the highest level of education from titles recorded in full name (e.g. "Mgr."). Can be used in conjunction with \code{msmt_addressess()}.
}
\examples{
data_addresses <- msmt_addresses(NUTS3 = "CZ010")
data_titles <- msmt_titles(data_addresses$ReditelJmeno)

}
