% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_1_utils.R
\name{set_utf8}
\alias{set_utf8}
\title{Set native encoding to UTF-8}
\usage{
set_utf8()
}
\description{
For the session duration, \code{set_utf8} changes native encoding to UTF-8,
in order to accommodate special Czech characters
}
